refactoring main page
